+++ panama-uring/src/main/java/top/dreamlike/panama/uring/eventloop/IoUringEventLoop.java
@@ -45,7 +45,7 @@ import java.util.function.Supplier;
 
 import static top.dreamlike.panama.uring.nativelib.struct.liburing.IoUringConstant.IORING_ASYNC_CANCEL_ALL;
 
-public sealed class IoUringEventLoop implements AutoCloseable, Executor, Runnable permits VTIoUringEventLoop, NettyEpollBridgeEventLoop {
+public sealed class IoUringEventLoop implements AutoCloseable, Executor, Runnable permits VTIoUringEventLoop, AbstractNettyBridgeEventLoop {
 
     protected static final AtomicInteger count = new AtomicInteger(0);
     protected static final LibUring libUring = Instance.LIB_URING;
